Completed: Order / Orders
Order / Order routes and page completed.

+++ src/app/api/categories/route.js
@@ -1,14 +1,19 @@
 import mongoose from 'mongoose';
 import { Category } from '../../../models/Category';
+import { isAdmin } from '../auth/[...nextauth]/route';
 
 export async function POST(req) {
 	// Connect to the DB:
 	mongoose.connect(process.env.MONGO_URL);
 
 	const { name } = await req.json();
-	// create the Category Model:
-	const categoryDoc = await Category.create({ name });
-	return Response.json(categoryDoc);
+	if (isAdmin()) {
+		// create the Category Model:
+		const categoryDoc = await Category.create({ name });
+		return Response.json(categoryDoc);
+	} else {
+		return Response.json({});
+	}
 }
 
 export async function PUT(req) {
@@ -16,7 +21,9 @@ export async function PUT(req) {
 	mongoose.connect(process.env.MONGO_URL);
 
 	const { _id, name } = await req.json();
-	await Category.updateOne({ _id }, { name });
+	if (await isAdmin()) {
+		await Category.updateOne({ _id }, { name });
+	}
 	return Response.json(true);
 }
 
@@ -30,9 +37,11 @@ export async function GET() {
 export async function DELETE(req) {
 	// connect to the DB:
 	mongoose.connect(process.env.MONGO_URL);
-
 	const url = new URL(req.url);
 	const _id = url.searchParams.get('_id');
-	await Category.deleteOne({ _id });
+
+	if (isAdmin()) {
+		await Category.deleteOne({ _id });
+	}
 	return Response.json(true);
 }
